// resources/js/types/family.d.ts
export type Family = {
  id: number;
  student_id: number;
  father_name: string;
  father_ocupation: string;
  father_address: string;
  father_phone: string;
  father_sallary: string;
  mother_name: string;
  mother_ocupation: string;
  mother_address: string;
  mother_phone: string;
  mother_sallary: string;
  created_at?: string;
  updated_at?: string;
};